* refactoring of re-usable code into fixtures.py. added a 'spacemake run -n' dry run test of the workflow logic
